stunnel-5.45

+++ src/fd.c
@@ -1,6 +1,6 @@
 /*
  *   stunnel       TLS offloading and load-balancing proxy
- *   Copyright (C) 1998-2017 Michal Trojnara <Michal.Trojnara@stunnel.org>
+ *   Copyright (C) 1998-2018 Michal Trojnara <Michal.Trojnara@stunnel.org>
  *
  *   This program is free software; you can redistribute it and/or modify it
  *   under the terms of the GNU General Public License as published by the
@@ -197,8 +197,8 @@ NOEXPORT SOCKET setup_fd(SOCKET fd, int nonblock, char *msg) {
     }
 #ifndef USE_FORK
     if(max_fds && fd>=max_fds) {
-        s_log(LOG_ERR, "%s: FD=%d out of range (max %d)",
-            msg, (int)fd, (int)max_fds);
+        s_log(LOG_ERR, "%s: FD=%ld out of range (max %d)",
+            msg, (long)fd, (int)max_fds);
         closesocket(fd);
         return INVALID_SOCKET;
     }
@@ -218,8 +218,8 @@ NOEXPORT SOCKET setup_fd(SOCKET fd, int nonblock, char *msg) {
 #endif /* USE_NEW_LINUX_API */
 
 #ifdef DEBUG_FD_ALLOC
-    s_log(LOG_DEBUG, "%s: FD=%d allocated (%sblocking mode)",
-        msg, fd, nonblock?"non-":"");
+    s_log(LOG_DEBUG, "%s: FD=%ld allocated (%sblocking mode)",
+        msg, (long)fd, nonblock?"non-":"");
 #endif /* DEBUG_FD_ALLOC */
 
     return fd;
